+++ packages/indexer/src/jobs/events-sync/index.ts
@@ -1,12 +1,11 @@
 import cron from "node-cron";
 
 import { logger } from "@/common/logger";
-import { safeWebSocketSubscription } from "@/common/provider";
+import { baseProvider, safeWebSocketSubscription } from "@/common/provider";
 import { redlock } from "@/common/redis";
 import { config } from "@/config/index";
 import { getNetworkSettings } from "@/config/network";
 import * as realtimeEventsSync from "@/jobs/events-sync/realtime-queue";
-import * as realtimeEventsSyncV2 from "@/jobs/events-sync/realtime-queue-v2";
 
 // For syncing events we have two separate job queues. One is for
 // handling backfilling of past event while the other one handles
@@ -44,7 +43,8 @@ if (config.doBackgroundWork && config.catchup) {
         )
         .then(async () => {
           try {
-            await realtimeEventsSync.addToQueue();
+            const block = await baseProvider.getBlockNumber();
+            await realtimeEventsSync.addToQueue({ block });
             logger.info("events-sync-catchup", "Catching up events");
           } catch (error) {
             logger.error("events-sync-catchup", `Failed to catch up events: ${error}`);
@@ -67,10 +67,7 @@ if (config.doBackgroundWork && config.catchup) {
         logger.info("events-sync-catchup", `Detected new block ${block}`);
 
         try {
-          await realtimeEventsSync.addToQueue();
-          if (config.enableRealtimeV2BlockQueue) {
-            await realtimeEventsSyncV2.addToQueue({ block });
-          }
+          await realtimeEventsSync.addToQueue({ block });
         } catch (error) {
           logger.error("events-sync-catchup", `Failed to catch up events: ${error}`);
         }
