+++ packages/indexer/src/models/blocks/index.ts
@@ -63,3 +63,29 @@ export const getBlocks = async (number: number): Promise<Block[]> =>
         timestamp,
       }))
     );
+
+export const getBlockWithNumber = async (number: number, hash: string): Promise<Block | null> =>
+  idb
+    .oneOrNone(
+      `
+        SELECT
+          blocks.hash,
+          blocks.timestamp
+        FROM blocks
+        WHERE blocks.number = $/number/
+          AND blocks.hash != $/hash/
+      `,
+      {
+        hash: toBuffer(hash),
+        number,
+      }
+    )
+    .then((result) =>
+      result
+        ? {
+            hash: fromBuffer(result.hash),
+            number,
+            timestamp: result.timestamp,
+          }
+        : null
+    );
